Move variableexplorer preferences to confpage.py
